requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cloudinary==1.30.0
cryptography==38.0.1
dj-database-url==1.0.0
Django==4.1.2
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-heroku==0.3.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
img2pdf==0.4.4
djangorestframework-simplejwt==5.2.2
django-filter==22.1
drf-extra-fields==3.4.1
gunicorn==20.1.0
idna==3.4
mysql-client==0.0.1
numpy==1.23.5
Pillow==9.2.0
pycodestyle==2.9.1
pycparser==2.21
psycopg2-binary==2.9.4
PyMySQL==1.0.2
python-decouple==3.6
pytz==2022.4
razorpay==1.3.0
requests==2.28.1
six==1.16.0
PyJWT==2.6.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.5
urllib3==1.26.12
whitenoise==6.2.0
python-dotenv==0.21.0
